finished commandbase
