+++ nextcord/ext/abc/_types.py
@@ -50,3 +50,4 @@ CommandT = TypeVar('CommandT', bound='CommandBase')
 ContextT = TypeVar('ContextT', bound='ContextBase')
 CoroFuncT = TypeVar('CoroFuncT', bound='CoroFunc')
 HookT = TypeVar('HookT', bound='Hook')
+ErrorT = TypeVar('ErrorT', bound='Error')
