+++ nextcord/ext/commands/core.py
@@ -124,35 +124,11 @@ CogT = TypeVar('CogT', bound='Cog')
 CommandT = TypeVar('CommandT', bound='Command')
 ContextT = TypeVar('ContextT', bound='Context')
 GroupT = TypeVar('GroupT', bound='Group')
-ErrorT = TypeVar('ErrorT', bound='Error')
 
 if TYPE_CHECKING:
     P = ParamSpec('P')
 else:
     P = TypeVar('P')
-
-
-def hooked_wrapped_callback(command, ctx, coro):
-    @functools.wraps(coro)
-    async def wrapped(*args, **kwargs):
-        try:
-            ret = await coro(*args, **kwargs)
-        except CommandError:
-            ctx.command_failed = True
-            raise
-        except asyncio.CancelledError:
-            ctx.command_failed = True
-            return
-        except Exception as exc:
-            ctx.command_failed = True
-            raise CommandInvokeError(exc) from exc
-        finally:
-            if command._max_concurrency is not None:
-                await command._max_concurrency.release(ctx)
-
-            await command.call_after_hooks(ctx)
-        return ret
-    return wrapped
 
 
 class _CaseInsensitiveDict(dict):
@@ -284,6 +260,8 @@ class Command(CommandBase):
         if not isinstance(self.aliases, (list, tuple)):
             raise TypeError("Aliases of a command must be a list or a tuple of strings.")
 
+        self.cooldown_after_parsing: bool = kwargs.get('cooldown_after_parsing', False)
+
         self.require_var_positional: bool = kwargs.get(
             'require_var_positional', False)
         self.ignore_extra: bool = kwargs.get('ignore_extra', True)
@@ -376,29 +354,6 @@ class Command(CommandBase):
             raise RuntimeError() from None  # break loop
         else:
             return value
-
-    @property
-    def clean_params(self) -> Dict[str, inspect.Parameter]:
-        """Dict[:class:`str`, :class:`inspect.Parameter`]:
-        Retrieves the parameter dictionary without the context or self parameters.
-
-        Useful for inspecting signature.
-        """
-        result = self.params.copy()
-        if self.cog is not None:
-            # first parameter is self
-            try:
-                del result[next(iter(result))]
-            except StopIteration:
-                raise ValueError("missing 'self' parameter") from None
-
-        try:
-            # first/second parameter is context
-            del result[next(iter(result))]
-        except StopIteration:
-            raise ValueError("missing 'context' parameter") from None
-
-        return result
 
     @property
     def full_parent_name(self) -> str:
@@ -517,220 +472,18 @@ class Command(CommandBase):
             raise TooManyArguments(
                 'Too many arguments passed to ' + self.qualified_name)
 
-    async def call_before_hooks(self, ctx: Context) -> None:
-        # now that we're done preparing we can call the pre-command hooks
-        # first, call the command local hook:
-        cog = self.cog
-        if self._before_invoke is not None:
-            # should be cog if @commands.before_invoke is used
-            instance = getattr(self._before_invoke, '__self__', cog)
-            # __self__ only exists for methods, not functions
-            # however, if @command.before_invoke is used, it will be a function
-            if instance:
-                await self._before_invoke(instance, ctx)  # type: ignore
-            else:
-                await self._before_invoke(ctx)  # type: ignore
+    async def _prepare_cooldowns(self, ctx: ContextT) -> None:
+        if self.cooldown_after_parsing:
+            await self._parse_arguments(ctx)
+            await super()._prepare_cooldowns(ctx)
+        else:
+            await super()._prepare_cooldowns(ctx)
+            await self._parse_arguments(ctx)
 
-        # call the cog local hook if applicable:
-        if cog is not None:
-            hook = Cog._get_overridden_method(cog.cog_before_invoke)
-            if hook is not None:
-                await hook(ctx)
-
-        # call the bot global hook if necessary
-        hook = ctx.bot._before_invoke
-        if hook is not None:
-            await hook(ctx)
-
-    async def call_after_hooks(self, ctx: Context) -> None:
-        cog = self.cog
-        if self._after_invoke is not None:
-            instance = getattr(self._after_invoke, '__self__', cog)
-            if instance:
-                await self._after_invoke(instance, ctx)  # type: ignore
-            else:
-                await self._after_invoke(ctx)  # type: ignore
-
-        # call the cog local hook if applicable:
-        if cog is not None:
-            hook = Cog._get_overridden_method(cog.cog_after_invoke)
-            if hook is not None:
-                await hook(ctx)
-
-        hook = ctx.bot._after_invoke
-        if hook is not None:
-            await hook(ctx)
-
-    def _prepare_cooldowns(self, ctx: Context) -> None:
-        if self._buckets.valid:
-            dt = ctx.message.edited_at or ctx.message.created_at
-            current = dt.replace(tzinfo=datetime.timezone.utc).timestamp()
-            bucket = self._buckets.get_bucket(ctx.message, current)
-            if bucket is not None:
-                retry_after = bucket.update_rate_limit(current)
-                if retry_after:
-                    raise CommandOnCooldown(
-                        bucket, retry_after, self._buckets.type)  # type: ignore
-
-    async def prepare(self, ctx: Context) -> None:
-        ctx.command = self
-
-        if not await self.can_run(ctx):
-            raise CheckFailure(
-                f'The check functions for command {self.qualified_name} failed.')
-
-        if self._max_concurrency is not None:
-            # For this application, context can be duck-typed as a Message
-            await self._max_concurrency.acquire(ctx)  # type: ignore
-
-        try:
-            if self.cooldown_after_parsing:
-                await self._parse_arguments(ctx)
-                self._prepare_cooldowns(ctx)
-            else:
-                self._prepare_cooldowns(ctx)
-                await self._parse_arguments(ctx)
-
-            await self.call_before_hooks(ctx)
-        except:
-            if self._max_concurrency is not None:
-                await self._max_concurrency.release(ctx)  # type: ignore
-            raise
-
-    def is_on_cooldown(self, ctx: Context) -> bool:
-        """Checks whether the command is currently on cooldown.
-
-        Parameters
-        -----------
-        ctx: :class:`.Context`
-            The invocation context to use when checking the commands cooldown status.
-
-        Returns
-        --------
-        :class:`bool`
-            A boolean indicating if the command is on cooldown.
-        """
-        if not self._buckets.valid:
-            return False
-
-        bucket = self._buckets.get_bucket(ctx.message)
-        dt = ctx.message.edited_at or ctx.message.created_at
-        current = dt.replace(tzinfo=datetime.timezone.utc).timestamp()
-        return bucket.get_tokens(current) == 0
-
-    def reset_cooldown(self, ctx: Context) -> None:
-        """Resets the cooldown on this command.
-
-        Parameters
-        -----------
-        ctx: :class:`.Context`
-            The invocation context to reset the cooldown under.
-        """
-        if self._buckets.valid:
-            bucket = self._buckets.get_bucket(ctx.message)
-            bucket.reset()
-
-    def get_cooldown_retry_after(self, ctx: Context) -> float:
-        """Retrieves the amount of seconds before this command can be tried again.
-
-        .. versionadded:: 1.4
-
-        Parameters
-        -----------
-        ctx: :class:`.Context`
-            The invocation context to retrieve the cooldown from.
-
-        Returns
-        --------
-        :class:`float`
-            The amount of time left on this command's cooldown in seconds.
-            If this is ``0.0`` then the command isn't on cooldown.
-        """
-        if self._buckets.valid:
-            bucket = self._buckets.get_bucket(ctx.message)
-            dt = ctx.message.edited_at or ctx.message.created_at
-            current = dt.replace(tzinfo=datetime.timezone.utc).timestamp()
-            return bucket.get_retry_after(current)
-
-        return 0.0
-
-    async def invoke(self, ctx: Context) -> None:
-        await self.prepare(ctx)
-
-        # terminate the invoked_subcommand chain.
-        # since we're in a regular command (and not a group) then
-        # the invoked subcommand is None.
-        ctx.invoked_subcommand = None
-        ctx.subcommand_passed = None
-        injected = hooked_wrapped_callback(self, ctx, self.callback)
-        await injected(*ctx.args, **ctx.kwargs)
-
-    async def reinvoke(self, ctx: Context, *, call_hooks: bool = False) -> None:
+    async def reinvoke(self, ctx: ContextT, *, call_hooks: bool = False) -> None:
         ctx.command = self
         await self._parse_arguments(ctx)
-
-        if call_hooks:
-            await self.call_before_hooks(ctx)
-
-        ctx.invoked_subcommand = None
-        try:
-            await self.callback(*ctx.args, **ctx.kwargs)  # type: ignore
-        except:
-            ctx.command_failed = True
-            raise
-        finally:
-            if call_hooks:
-                await self.call_after_hooks(ctx)
-
-    def error(self, coro: ErrorT) -> ErrorT:
-        """A decorator that registers a coroutine as a local error handler.
-
-        A local error handler is an :func:`.on_command_error` event limited to
-        a single command. However, the :func:`.on_command_error` is still
-        invoked afterwards as the catch-all.
-
-        Parameters
-        -----------
-        coro: :ref:`coroutine <coroutine>`
-            The coroutine to register as the local error handler.
-
-        Raises
-        -------
-        TypeError
-            The coroutine passed is not actually a coroutine.
-        """
-
-        if not asyncio.iscoroutinefunction(coro):
-            raise TypeError('The error handler must be a coroutine.')
-
-        self.on_error: Error = coro
-        return coro
-
-    def has_error_handler(self) -> bool:
-        """:class:`bool`: Checks whether the command has an error handler registered.
-
-        .. versionadded:: 1.7
-        """
-        return hasattr(self, 'on_error')
-
-    @property
-    def cog_name(self) -> Optional[str]:
-        """Optional[:class:`str`]: The name of the cog this command belongs to, if any."""
-        return type(self.cog).__cog_name__ if self.cog is not None else None
-
-    @property
-    def short_doc(self) -> str:
-        """:class:`str`: Gets the "short" documentation of a command.
-
-        By default, this is the :attr:`.brief` attribute.
-        If that lookup leads to an empty string then the first line of the
-        :attr:`.help` attribute is used instead.
-        """
-        if self.brief is not None:
-            return self.brief
-        if self.help is not None:
-            return self.help.split('\n', 1)[0]
-        return ''
+        super().reinvoke(ctx, call_hooks=call_hooks)
 
     def _is_typing_optional(self, annotation: Union[T, Optional[T]]) -> TypeGuard[Optional[T]]:
         # type: ignore
@@ -791,62 +544,6 @@ class Command(CommandBase):
                 result.append(f'<{name}>')
 
         return ' '.join(result)
-
-    async def can_run(self, ctx: Context) -> bool:
-        """|coro|
-
-        Checks if the command can be executed by checking all the predicates
-        inside the :attr:`~Command.checks` attribute. This also checks whether the
-        command is disabled.
-
-        .. versionchanged:: 1.3
-            Checks whether the command is disabled or not
-
-        Parameters
-        -----------
-        ctx: :class:`.Context`
-            The ctx of the command currently being invoked.
-
-        Raises
-        -------
-        :class:`CommandError`
-            Any command error that was raised during a check call will be propagated
-            by this function.
-
-        Returns
-        --------
-        :class:`bool`
-            A boolean indicating if the command can be invoked.
-        """
-
-        if not self.enabled:
-            raise DisabledCommand(f'{self.name} command is disabled')
-
-        original = ctx.command
-        ctx.command = self
-
-        try:
-            if not await ctx.bot.can_run(ctx):
-                raise CheckFailure(
-                    f'The global check functions for command {self.qualified_name} failed.')
-
-            cog = self.cog
-            if cog is not None:
-                local_check = Cog._get_overridden_method(cog.cog_check)
-                if local_check is not None:
-                    ret = await nextcord.utils.maybe_coroutine(local_check, ctx)
-                    if not ret:
-                        return False
-
-            predicates = self.checks
-            if not predicates:
-                # since we have no checks, then we just return True.
-                return True
-
-            # type: ignore
-            return await nextcord.utils.async_all(predicate(ctx) for predicate in predicates)
-        finally:
-            ctx.command = original
 
 
 class GroupMixin(Generic[CogT]):
@@ -1902,13 +1599,13 @@ def cooldown(rate: int, per: float, type: Union[BucketType, Callable[[Message], 
         .. versionchanged:: 1.7
             Callables are now supported for custom bucket types.
     """
+    wrapped: Union[BucketType, Callable[[ContextT], Any]] = type if isinstance(type, BucketType) else lambda ctx: type(ctx.message)
 
     def decorator(func: Union[Command, CoroFunc]) -> Union[Command, CoroFunc]:
         if isinstance(func, Command):
-            func._buckets = CooldownMapping(Cooldown(rate, per), type)
+            func._buckets = CooldownMapping(Cooldown(rate, per), wrapped)
         else:
-            func.__commands_cooldown__ = CooldownMapping(
-                Cooldown(rate, per), type)
+            func.__commands_cooldown__ = CooldownMapping(Cooldown(rate, per), wrapped)
         return func
     return decorator  # type: ignore
 
@@ -1945,11 +1642,12 @@ def dynamic_cooldown(cooldown: Union[BucketType, Callable[[Message], Any]], type
     if not callable(cooldown):
         raise TypeError("A callable must be provided")
 
+    wrapped = lambda ctx: cooldown(ctx.message)
     def decorator(func: Union[Command, CoroFunc]) -> Union[Command, CoroFunc]:
         if isinstance(func, Command):
-            func._buckets = DynamicCooldownMapping(cooldown, type)
+            func._buckets = DynamicCooldownMapping(wrapped, type)
         else:
-            func.__commands_cooldown__ = DynamicCooldownMapping(cooldown, type)
+            func.__commands_cooldown__ = DynamicCooldownMapping(wrapped, type)
         return func
     return decorator  # type: ignore
